+++ test/test_SelectList.py
@@ -8,7 +8,6 @@ class testSL(SeleniumUtils, SetupByLocation):
 
     def setUp(self):
         super(testSL,self).setUp()
-        self.driver.implicitly_wait(30) # seconds
         self.url='http://'+self.IP+'/examples/QuickStartSelect.html'
 
     def test02_chk_inp(self):
@@ -25,10 +24,10 @@ class testSL(SeleniumUtils, SetupByLocation):
            that any character not in select list is deleted
         """
         self.open_n_tst_title({'url': self.url, 'title': 'MenuOptions'} )
-        self.check_clr({ 'xpath': '//*[@id="CB_menuoptions3"]',
+        self.check_clr({ 'xpath': '//*[@id="CB_menuoptions4"]',
                          'input': '//*[@id="selecttest"]' })
         self.check_autocomplete({ 'xpath': '//*[@id="selecttest"]',
-                         'filt_rslts': '//*[@id="SP_menuoptions3"]',
+                         'filt_rslts': '//*[@id="SP_menuoptions4"]',
                          'expected': 'JanuaryFebruaryMarchAprilMayAugust',
                          'test_key': 'a' })
         self.check_invalid_key ({ 'xpath': '//*[@id="selecttest"]',
@@ -38,10 +37,10 @@ class testSL(SeleniumUtils, SetupByLocation):
         """
            check that MenuOptions select list autocomplete works with images
         """
-        self.url='http://www.menuoptions.org/examples/SelectWithImages.html'
+        self.url='http://'+self.IP+'/examples/SelectWithImages.html'
         self.open_n_tst_title({'url': self.url, 'title': 'Select'} )
         self.check_autocomplete({ 'xpath': '//*[@id="CustSel"]',
-                         'filt_rslts': '//*[@id="SP_menuoptions3"]',
+                         'filt_rslts': '//*[@id="SP_menuoptions4"]',
                          'expected': 'CashChargeCheck',
                          'test_key': 'c' })