+++ test/test_addons.py
@@ -10,7 +10,6 @@ class test_addons(SetupByLocation, SeleniumUtils):
 
     def setUp(self):
         super(test_addons,self).setUp()
-        self.driver.implicitly_wait(30) # seconds
         self.url='http://'+self.IP+'/examples/QuickStartSelect.html'
 
     def test02_coloring(self):
@@ -19,19 +18,33 @@ class test_addons(SetupByLocation, SeleniumUtils):
         """
         self.open_n_tst_title({'url': self.url, 'title': 'MenuOptions'} )
         self.check_html({ 'xpath': '//*[@id="scrolltest"]',
-                             'cell': '//*[@id="SP_menuoptions4"]/table/tbody/tr/td',
+                             'cell': '//*[@id="SP_menuoptions5"]/table/tbody/tr/td',
                              'expected': '<span style="color:brown;font-size:110%;">December</span>'})
 
-    def test03_rt_arrow(self):
+    def test03_down_arrow(self):
+        """
+           verify that  dropdown menus show a down arrow after menu text
+        """
+        self.url='http://'+self.IP+'/examples/MenusBottom.html'
+        self.open_n_tst_title({'url': self.url, 'title': 'MenuOptions'} )
+        self.check_html({ 'xpath': '//*[@id="menu_plain"]/span',
+            'partial': True,
+            'expected': u'Menu&nbsp;.*span.*class="down_arrow"' })
+        self.check_html({ 'xpath': '//*[@id="menu_w_imgs"]/span',
+            'partial': True,
+            'expected': u'Menuwithimages&nbsp;.*span.*class="down_arrow"' })
+    def test04_rt_arrow(self):
         """
            verify that right side menus show a right arrow after menu text
         """
         self.url='http://'+self.IP+'/examples/RightMenu.html'
         self.open_n_tst_title({'url': self.url, 'title': 'MenuOptions'} )
         self.check_html({ 'xpath': '//*[@id="menu_plain"]/span',
-                          'expected': u'Menu&nbsp;▸'.encode('utf-8')})
+            'partial': True,
+            'expected': u'Menu&nbsp;.*span.*class="right_arrow"' })
         self.check_html({ 'xpath': '//*[@id="menu_w_imgs"]/span',
-                          'expected': u'Menuwithimages&nbsp;▸'.encode('utf-8')})
+            'partial': True,
+            'expected': u'Menuwithimages&nbsp;.*span.*class="right_arrow"' })
 
     def tearDown(self):
         super(test_addons,self).tearDown()